+++ controllers/pages/events.js
@@ -2,10 +2,10 @@
 
 // Imports
 const AWS = require('aws-sdk')
-const trips = require('../data_managers/trips.js')
+const trips = require('../../data_managers/trips.js')
 const s3 = new AWS.S3()
-const logger = require('../log.js')
-const viewData = require('../view_data.js')
+const logger = require('../../log.js')
+const viewData = require('../../view_data.js')
 
 const events = {
     async get(req, res) {
